Add convenince methods to response class
Surface the details a developer will want to query when getting a response.
