+++ lib/ibm/cloud/sdk/vpc/helpers/connection.rb
@@ -2,6 +2,7 @@
 
 require 'time'
 require 'http'
+require_relative 'response'
 
 module IBM
   module Cloud
@@ -57,44 +58,6 @@ module IBM
           def authorization_header
             fetch if expired?
             "#{@data.fetch(:token_type)} #{@data.fetch(:access_token)}"
-          end
-        end
-
-        class Response
-
-          def initialize(response)
-            @response = response
-          end
-
-          attr_reader :response
-
-          def json
-            JSON.parse(response.body, symbolize_names: true)
-          rescue StandardError
-            raise "Error while parsing response body. #{response.body}"
-          end
-
-          def hash_response
-            check_object(Hash)
-          end
-
-          def array_rasponse
-            check_object(Array)
-          end
-
-          def subkey(key)
-            ret = hash_response
-            sym_key = key.to_sym
-            return ret.fetch(sym_key) if ret.key?(sym_key)
-
-            raise "Key not found in #{ret}."
-          end
-
-          def check_object(obj)
-            ret = json
-            return ret if ret.instance_of?(obj)
-
-            raise "Expected #{obj} in response not #{ret.class}."
           end
         end
       end
